homework2_app/requirements.txt
flask
gunicorn
cryptography
flask-socketio
gevent>=23.0.0
gevent-websocket
dnspython>=2.0.0
google-generativeai
groq
psycopg2-binary
requests
python-dotenv
beautifulsoup4
Flask-Failsafe
pgvector